Улучшена структура проекта
Resolve https://github.com/stas12312/my-tickets-bot/issues/147
