+++ my_tickets_bot/src/bot/callbacks/action.py
@@ -1,0 +1,17 @@
+from enum import StrEnum
+
+from aiogram.filters.callback_data import CallbackData
+
+
+class CloseCallback(CallbackData, prefix='close'):
+    """CallbackData для закрытия меню"""
+
+
+class EntityAction(StrEnum):
+    """Действия с сущностями"""
+    ADD = 'add'
+    UPDATE = 'update'
+    DELETE = 'delete'
+    LIST = 'list'
+    SHOW = 'show'
+    EDIT = 'edit'
