+++ my_tickets_bot/src/bot/callbacks/pagination.py
@@ -1,0 +1,7 @@
+from aiogram.filters.callback_data import CallbackData
+
+
+class PaginationCallback(CallbackData, prefix='page'):
+    """CallbackData для пагинации"""
+    object_name: str
+    page: int | None
