+++ my_tickets_bot/src/bot/forms/__init__.py
@@ -1,0 +1,4 @@
+from .city import CityForm
+from .event import EventForm, EditEventForm
+from .location import LocationForm
+from .ticket import TicketForm
