+++ my_tickets_bot/src/bot/forms/city.py
@@ -1,0 +1,7 @@
+from aiogram.fsm.state import StatesGroup, State
+
+
+class CityForm(StatesGroup):
+    """Добавление города"""
+    name = State()
+    timezone = State()
