+++ my_tickets_bot/src/bot/forms/location.py
@@ -1,0 +1,8 @@
+from aiogram.fsm.state import StatesGroup, State
+
+
+class LocationForm(StatesGroup):
+    """Добавление места"""
+    input_name = State()
+    input_address = State()
+    input_url = State()
