+++ my_tickets_bot/src/bot/forms/ticket.py
@@ -1,0 +1,6 @@
+from aiogram.fsm.state import StatesGroup, State
+
+
+class TicketForm(StatesGroup):
+    """Добавление билета"""
+    file = State()
