+++ my_tickets_bot/src/bot/keyboards/city.py
@@ -3,7 +3,7 @@ from aiogram.utils.keyboard import InlineKeyboardBuilder
 
 from bot.buttons import Settings
 from bot.callbacks import CityCallback, EntityAction, SettingsCallback, LocationCallback
-from bot.keybaords import CLOSE_BUTTON
+from bot.keyboards.utils import CLOSE_BUTTON
 from models import City
