+++ my_tickets_bot/src/bot/messages/statistic.py
@@ -1,29 +1,6 @@
-"""Формирование красивых сообщений"""
-
 from aiogram.utils.markdown import bold
-from aiogram.utils.text_decorations import markdown_decoration
 
 from services.statistic import RowStatistic
-
-
-def quote(
-        value: str
-) -> str:
-    """Экранирование"""
-    return markdown_decoration.quote(value)
-
-
-TIME_EXAMPLES = f'Примеры:\n' \
-                f'_{quote("20.03.23 20:00")}_\n' \
-                f'_{quote("20.03 19:00")}_\n' \
-                f'_{quote("20 марта 21:30")}_'
-
-
-def _make_message_by_rows(
-        rows: list[str],
-) -> str:
-    """Формирование сообщения из списка строк"""
-    return '\n'.join(rows)
 
 
 def get_message_for_statistic(
