+++ my_tickets_bot/src/bot/messages/templates.py
@@ -1,0 +1,6 @@
+from .utils import quote
+
+TIME_EXAMPLES = f'Примеры:\n' \
+                f'_{quote("20.03.23 20:00")}_\n' \
+                f'_{quote("20.03 19:00")}_\n' \
+                f'_{quote("20 марта 21:30")}_'
